Make SymbolRenamer public (#1161)
