+++ src/CommandLine/Program.cs
@@ -16,6 +16,7 @@ using CommandLine.Text;
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CodeFixes;
 using Roslynator.CodeFixes;
+using Roslynator.CommandLine.Rename;
 using Roslynator.Diagnostics;
 using Roslynator.Documentation;
 using Roslynator.FindSymbols;
@@ -430,18 +431,10 @@ internal static class Program
         if (!TryParsePaths(options.Paths, out ImmutableArray<string> paths))
             return ExitCodes.Error;
 
-        if (!TryParseOptionValueAsEnum(options.OnError, OptionNames.OnError, out RenameErrorResolution errorResolution, defaultValue: RenameErrorResolution.None))
+        if (!TryParseOptionValueAsEnum(options.OnError, OptionNames.OnError, out CliCompilationErrorResolution errorResolution, defaultValue: CliCompilationErrorResolution.None))
             return ExitCodes.Error;
 
-#pragma warning disable RCS1118
-        var visibility = Visibility.Public;
-#pragma warning restore RCS1118
-        var scopeFilter = RenameScopeFilter.All;
-#if DEBUG
-        if (!TryParseOptionValueAsEnum(options.Visibility, OptionNames.Visibility, out visibility))
-            return ExitCodes.Error;
-#endif
-        if (!TryParseOptionValueAsEnumFlags(options.Scope, OptionNames.Scope, out scopeFilter, defaultValue: RenameScopeFilter.All))
+        if (!TryParseOptionValueAsEnumFlags(options.Scope, OptionNames.Scope, out RenameScopeFilter scopeFilter, defaultValue: RenameScopeFilter.All))
             return ExitCodes.Error;
 
         if (!TryParseCodeExpression(
@@ -486,15 +479,9 @@ internal static class Program
             options: options,
             projectFilter: projectFilter,
             scopeFilter: scopeFilter,
-            visibility: visibility,
             errorResolution: errorResolution,
-#if DEBUG
             ignoredCompilerDiagnostics: options.IgnoredCompilerDiagnostics,
-            codeContext: options.CodeContext,
-#else
-            ignoredCompilerDiagnostics: null,
             codeContext: -1,
-#endif
             predicate: predicate,
             getNewName: getNewName);
