+++ src/Workspaces.Core/CodeFixes/CodeFixerOptions.cs
@@ -39,8 +39,13 @@ internal class CodeFixerOptions : CodeAnalysisOptions
             ignoredDiagnosticIds: ignoredDiagnosticIds)
     {
         IgnoreCompilerErrors = ignoreCompilerErrors;
-        IgnoredCompilerDiagnosticIds = ignoredCompilerDiagnosticIds?.ToImmutableHashSet() ?? ImmutableHashSet<string>.Empty;
         DiagnosticIdsFixableOneByOne = diagnosticIdsFixableOneByOne?.ToImmutableHashSet() ?? ImmutableHashSet<string>.Empty;
+
+        if (ignoredCompilerDiagnosticIds is not null)
+        {
+            foreach (string id in ignoredCompilerDiagnosticIds)
+                IgnoredCompilerDiagnosticIds.Add(id);
+        }
 
         if (diagnosticFixMap is not null)
         {
@@ -70,7 +75,7 @@ internal class CodeFixerOptions : CodeAnalysisOptions
 
     public bool IgnoreCompilerErrors { get; }
 
-    public ImmutableHashSet<string> IgnoredCompilerDiagnosticIds { get; }
+    public HashSet<string> IgnoredCompilerDiagnosticIds { get; } = new();
 
     public string FileBanner { get; }
