+++ tools/generate_api_list.ps1
@@ -1,0 +1,6 @@
+roslynator list-symbols generate_ref_docs.sln `
+ --properties Configuration=Release `
+ --visibility public `
+ --depth member `
+ --ignored-parts containing-namespace assembly-attributes `
+ --output "api.txt"
